requirements.txt
flask==3.1.0
openpyxl==3.1.5
Pillow==11.3.0
gunicorn==23.0.0
